requirements.txt
flask
flask_sqlalchemy
pymysql
flask-migrate
flask-wtf
email_validator